chore: wrap heavy routes with suspense
